+++ centreon/packages/ui/src/Form/Inputs/models.ts
@@ -60,6 +60,7 @@ export interface InputProps {
     filterKey?: string;
     getRenderedOptionText?: (option) => string | JSX.Element;
     disableSelectAll?: boolean;
+    limitTags?: number;
   };
   file?: {
     multiple?: boolean;
